relay/pyproject.toml
[project]
name = "litellm-o3-deep-research-relay"
version = "0.1.0"
description = "FastAPI + Hypercorn relay example that wraps the LiteLLM Python SDK"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.128,<1",
    "hypercorn>=0.18,<1",
    "litellm>=1.81.13,<2",
    "python-dotenv>=1,<2",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28,<1",
    "pytest>=8,<9",
    "pytest-asyncio>=0.23,<1",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
